Added shared structures for win and linux stat query
